refactor: split server stats connection helpers
