+++ src/backend/ssh/server-stats.ts
@@ -6,7 +6,7 @@ import { Client, type ConnectConfig } from "ssh2";
 import { SSH_ALGORITHMS } from "../utils/ssh-algorithms.js";
 import { getDb } from "../database/db/index.js";
 import { hosts, sshCredentials } from "../database/db/schema.js";
-import { eq, and } from "drizzle-orm";
+import { eq } from "drizzle-orm";
 import { statsLogger } from "../utils/logger.js";
 import { SimpleDBOps } from "../utils/simple-db-ops.js";
 import { AuthManager } from "../utils/auth-manager.js";
@@ -31,6 +31,13 @@ import { SSHHostKeyVerifier } from "./host-key-verifier.js";
 import { connectionPool, withConnection } from "./ssh-connection-pool.js";
 import { registerServerStatsSettingsRoutes } from "./server-stats-settings-routes.js";
 import { registerServerStatsViewerRoutes } from "./server-stats-viewer-routes.js";
+import { createJumpHostChain } from "./server-stats-jump-hosts.js";
+import {
+  authFailureTracker,
+  metricsCache,
+  pollingBackoff,
+  requestQueue,
+} from "./server-stats-state.js";
 
 function supportsMetrics(host: SSHHostWithCredentials): boolean {
   const connectionType = host.connectionType || "ssh";
@@ -55,319 +62,6 @@ function createConnectionLog(
     message,
     details,
   };
-}
-
-interface JumpHostConfig {
-  id: number;
-  ip: string;
-  port: number;
-  username: string;
-  password?: string;
-  key?: string;
-  keyPassword?: string;
-  keyType?: string;
-  authType?: string;
-  credentialId?: number;
-  [key: string]: unknown;
-}
-
-async function resolveJumpHost(
-  hostId: number,
-  userId: string,
-): Promise<JumpHostConfig | null> {
-  try {
-    const hostResults = await SimpleDBOps.select(
-      getDb().select().from(hosts).where(eq(hosts.id, hostId)),
-      "ssh_data",
-      userId,
-    );
-
-    if (hostResults.length === 0) {
-      return null;
-    }
-
-    const host = hostResults[0];
-    const ownerId = (host.userId || userId) as string;
-
-    if (host.credentialId) {
-      if (userId !== ownerId) {
-        try {
-          const { SharedCredentialManager } =
-            await import("../utils/shared-credential-manager.js");
-          const sharedCredManager = SharedCredentialManager.getInstance();
-          const sharedCred = await sharedCredManager.getSharedCredentialForUser(
-            hostId,
-            userId,
-          );
-          if (sharedCred) {
-            return {
-              ...host,
-              password: sharedCred.password,
-              key: sharedCred.key,
-              keyPassword: sharedCred.keyPassword,
-              keyType: sharedCred.keyType,
-              authType: sharedCred.key
-                ? "key"
-                : sharedCred.password
-                  ? "password"
-                  : "none",
-            } as JumpHostConfig;
-          }
-        } catch {
-          // fall through to owner credential lookup
-        }
-      }
-
-      const credentials = await SimpleDBOps.select(
-        getDb()
-          .select()
-          .from(sshCredentials)
-          .where(
-            and(
-              eq(sshCredentials.id, host.credentialId as number),
-              eq(sshCredentials.userId, ownerId),
-            ),
-          ),
-        "ssh_credentials",
-        ownerId,
-      );
-
-      if (credentials.length > 0) {
-        const credential = credentials[0];
-        return {
-          ...host,
-          password: credential.password as string | undefined,
-          key: (credential.key || credential.privateKey) as string | undefined,
-          keyPassword: credential.keyPassword as string | undefined,
-          keyType: credential.keyType as string | undefined,
-          authType: credential.authType as string | undefined,
-        } as JumpHostConfig;
-      }
-    }
-
-    return host as JumpHostConfig;
-  } catch (error) {
-    statsLogger.error("Failed to resolve jump host", error, {
-      operation: "resolve_jump_host",
-      hostId,
-      userId,
-    });
-    return null;
-  }
-}
-
-async function createJumpHostChain(
-  jumpHosts: Array<{ hostId: number }>,
-  userId: string,
-  socks5Config?: SOCKS5Config | null,
-): Promise<Client | null> {
-  if (!jumpHosts || jumpHosts.length === 0) {
-    return null;
-  }
-
-  let currentClient: Client | null = null;
-  const clients: Client[] = [];
-
-  try {
-    const jumpHostConfigs: Array<Awaited<ReturnType<typeof resolveJumpHost>>> =
-      [];
-    for (let i = 0; i < jumpHosts.length; i++) {
-      const config = await resolveJumpHost(jumpHosts[i].hostId, userId);
-      jumpHostConfigs.push(config);
-    }
-
-    const totalHops = jumpHostConfigs.length;
-
-    for (let i = 0; i < jumpHostConfigs.length; i++) {
-      if (!jumpHostConfigs[i]) {
-        statsLogger.error(`Jump host ${i + 1} not found`, undefined, {
-          operation: "jump_host_chain",
-          hostId: jumpHosts[i].hostId,
-          hopIndex: i,
-          totalHops,
-        });
-        clients.forEach((c) => c.end());
-        return null;
-      }
-    }
-
-    let proxySocket: import("net").Socket | null = null;
-    if (socks5Config?.useSocks5) {
-      const firstHop = jumpHostConfigs[0]!;
-      proxySocket = await createSocks5Connection(
-        firstHop.ip,
-        firstHop.port || 22,
-        socks5Config,
-      );
-    }
-
-    for (let i = 0; i < jumpHostConfigs.length; i++) {
-      const jumpHostConfig = jumpHostConfigs[i]!;
-
-      const jumpClient = new Client();
-      clients.push(jumpClient);
-
-      const jumpHostVerifier = await SSHHostKeyVerifier.createHostVerifier(
-        jumpHostConfig.id,
-        jumpHostConfig.ip,
-        jumpHostConfig.port || 22,
-        null,
-        userId,
-        true,
-      );
-
-      const connected = await new Promise<boolean>((resolve) => {
-        const timeout = setTimeout(() => {
-          resolve(false);
-        }, 30000);
-
-        jumpClient.on("ready", () => {
-          clearTimeout(timeout);
-          resolve(true);
-        });
-
-        jumpClient.on("error", (err) => {
-          clearTimeout(timeout);
-          statsLogger.error(
-            `Jump host ${i + 1}/${totalHops} connection failed`,
-            err,
-            {
-              operation: "jump_host_connect",
-              hostId: jumpHostConfig.id,
-              ip: jumpHostConfig.ip,
-              hopIndex: i,
-              totalHops,
-              previousHop:
-                i > 0
-                  ? jumpHostConfigs[i - 1]?.ip
-                  : proxySocket
-                    ? "proxy"
-                    : "direct",
-              usedProxySocket: i === 0 && !!proxySocket,
-            },
-          );
-          resolve(false);
-        });
-
-        const connectConfig: Record<string, unknown> = {
-          host: jumpHostConfig.ip?.replace(/^\[|\]$/g, "") || jumpHostConfig.ip,
-          port: jumpHostConfig.port || 22,
-          username: jumpHostConfig.username,
-          tryKeyboard: jumpHostConfig.authType !== "none",
-          readyTimeout: 60000,
-          hostVerifier: jumpHostVerifier,
-          algorithms: {
-            kex: [
-              "curve25519-sha256",
-              "curve25519-sha256@libssh.org",
-              "ecdh-sha2-nistp521",
-              "ecdh-sha2-nistp384",
-              "ecdh-sha2-nistp256",
-              "diffie-hellman-group-exchange-sha256",
-              "diffie-hellman-group18-sha512",
-              "diffie-hellman-group17-sha512",
-              "diffie-hellman-group16-sha512",
-              "diffie-hellman-group15-sha512",
-              "diffie-hellman-group14-sha256",
-              "diffie-hellman-group14-sha1",
-              "diffie-hellman-group-exchange-sha1",
-              "diffie-hellman-group1-sha1",
-            ],
-            serverHostKey: [
-              "ssh-ed25519",
-              "ecdsa-sha2-nistp521",
-              "ecdsa-sha2-nistp384",
-              "ecdsa-sha2-nistp256",
-              "rsa-sha2-512",
-              "rsa-sha2-256",
-              "ssh-rsa",
-              "ssh-dss",
-            ],
-            cipher: SSH_ALGORITHMS.cipher,
-            hmac: [
-              "hmac-sha2-512-etm@openssh.com",
-              "hmac-sha2-256-etm@openssh.com",
-              "hmac-sha2-512",
-              "hmac-sha2-256",
-              "hmac-sha1",
-              "hmac-md5",
-            ],
-            compress: ["none", "zlib@openssh.com", "zlib"],
-          },
-        };
-
-        if (jumpHostConfig.authType === "password" && jumpHostConfig.password) {
-          connectConfig.password = jumpHostConfig.password;
-        } else if (jumpHostConfig.authType === "key" && jumpHostConfig.key) {
-          const cleanKey = jumpHostConfig.key
-            .trim()
-            .replace(/\r\n/g, "\n")
-            .replace(/\r/g, "\n");
-          connectConfig.privateKey = Buffer.from(cleanKey, "utf8");
-          if (jumpHostConfig.keyPassword) {
-            connectConfig.passphrase = jumpHostConfig.keyPassword;
-          }
-        }
-
-        jumpClient.on(
-          "keyboard-interactive",
-          (
-            _name: string,
-            _instructions: string,
-            _lang: string,
-            prompts: Array<{ prompt: string; echo: boolean }>,
-            finish: (responses: string[]) => void,
-          ) => {
-            const responses = prompts.map((p) => {
-              if (/password/i.test(p.prompt) && jumpHostConfig.password) {
-                return jumpHostConfig.password as string;
-              }
-              return "";
-            });
-            finish(responses);
-          },
-        );
-
-        if (currentClient) {
-          currentClient.forwardOut(
-            "127.0.0.1",
-            0,
-            jumpHostConfig.ip,
-            jumpHostConfig.port || 22,
-            (err, stream) => {
-              if (err) {
-                clearTimeout(timeout);
-                resolve(false);
-                return;
-              }
-              connectConfig.sock = stream;
-              jumpClient.connect(connectConfig);
-            },
-          );
-        } else if (proxySocket) {
-          connectConfig.sock = proxySocket;
-          jumpClient.connect(connectConfig);
-        } else {
-          jumpClient.connect(connectConfig);
-        }
-      });
-
-      if (!connected) {
-        clients.forEach((c) => c.end());
-        return null;
-      }
-
-      currentClient = jumpClient;
-    }
-
-    return currentClient;
-  } catch (error) {
-    statsLogger.error("Failed to create jump host chain", error, {
-      operation: "jump_host_chain",
-    });
-    clients.forEach((c) => c.end());
-    return null;
-  }
 }
 
 interface MetricsSession {
@@ -448,262 +142,6 @@ function getSessionKey(hostId: number, userId: string): string {
   return `${userId}:${hostId}`;
 }
 
-class RequestQueue {
-  private queues = new Map<number, Array<() => Promise<unknown>>>();
-  private processing = new Set<number>();
-  private requestTimeout = 60000;
-
-  async queueRequest<T>(hostId: number, request: () => Promise<T>): Promise<T> {
-    return new Promise<T>((resolve, reject) => {
-      const wrappedRequest = async () => {
-        try {
-          const result = await Promise.race<T>([
-            request(),
-            new Promise<never>((_, rej) =>
-              setTimeout(
-                () =>
-                  rej(
-                    new Error(
-                      `Request timeout after ${this.requestTimeout}ms for host ${hostId}`,
-                    ),
-                  ),
-                this.requestTimeout,
-              ),
-            ),
-          ]);
-          resolve(result);
-        } catch (error) {
-          reject(error);
-        }
-      };
-
-      const queue = this.queues.get(hostId) || [];
-      queue.push(wrappedRequest);
-      this.queues.set(hostId, queue);
-      this.processQueue(hostId);
-    });
-  }
-
-  private async processQueue(hostId: number): Promise<void> {
-    if (this.processing.has(hostId)) return;
-
-    this.processing.add(hostId);
-    const queue = this.queues.get(hostId) || [];
-
-    while (queue.length > 0) {
-      const request = queue.shift();
-      if (request) {
-        try {
-          await request();
-        } catch {
-          // expected
-        }
-      }
-    }
-
-    this.processing.delete(hostId);
-    const currentQueue = this.queues.get(hostId);
-    if (currentQueue && currentQueue.length > 0) {
-      this.processQueue(hostId);
-    }
-  }
-}
-
-interface CachedMetrics {
-  data: unknown;
-  timestamp: number;
-  hostId: number;
-}
-
-class MetricsCache {
-  private cache = new Map<number, CachedMetrics>();
-  private ttl = 30000;
-
-  get(hostId: number): unknown | null {
-    const cached = this.cache.get(hostId);
-    if (cached && Date.now() - cached.timestamp < this.ttl) {
-      return cached.data;
-    }
-    return null;
-  }
-
-  set(hostId: number, data: unknown): void {
-    this.cache.set(hostId, {
-      data,
-      timestamp: Date.now(),
-      hostId,
-    });
-  }
-
-  clear(hostId?: number): void {
-    if (hostId) {
-      this.cache.delete(hostId);
-    } else {
-      this.cache.clear();
-    }
-  }
-}
-
-interface AuthFailureRecord {
-  count: number;
-  lastFailure: number;
-  reason: "TOTP" | "AUTH" | "TIMEOUT";
-  permanent: boolean;
-}
-
-class AuthFailureTracker {
-  private failures = new Map<number, AuthFailureRecord>();
-  private maxRetries = 3;
-  private backoffBase = 5000;
-
-  recordFailure(
-    hostId: number,
-    reason: "TOTP" | "AUTH" | "TIMEOUT",
-    permanent = false,
-  ): void {
-    const existing = this.failures.get(hostId);
-    if (existing) {
-      existing.count++;
-      existing.lastFailure = Date.now();
-      existing.reason = reason;
-      if (permanent) existing.permanent = true;
-    } else {
-      this.failures.set(hostId, {
-        count: 1,
-        lastFailure: Date.now(),
-        reason,
-        permanent,
-      });
-    }
-  }
-
-  shouldSkip(hostId: number): boolean {
-    const record = this.failures.get(hostId);
-    if (!record) return false;
-
-    if (record.reason === "TOTP" || record.permanent) {
-      return true;
-    }
-
-    if (record.count >= this.maxRetries) {
-      return true;
-    }
-
-    const backoffTime = this.backoffBase * Math.pow(2, record.count - 1);
-    const timeSinceFailure = Date.now() - record.lastFailure;
-
-    return timeSinceFailure < backoffTime;
-  }
-
-  getSkipReason(hostId: number): string | null {
-    const record = this.failures.get(hostId);
-    if (!record) return null;
-
-    if (record.reason === "TOTP") {
-      return "TOTP authentication required (metrics unavailable)";
-    }
-
-    if (record.permanent) {
-      return "Authentication permanently failed";
-    }
-
-    if (record.count >= this.maxRetries) {
-      return `Too many authentication failures (${record.count} attempts)`;
-    }
-
-    const backoffTime = this.backoffBase * Math.pow(2, record.count - 1);
-    const timeSinceFailure = Date.now() - record.lastFailure;
-    const remainingTime = Math.ceil((backoffTime - timeSinceFailure) / 1000);
-
-    if (timeSinceFailure < backoffTime) {
-      return `Retry in ${remainingTime}s (attempt ${record.count}/${this.maxRetries})`;
-    }
-
-    return null;
-  }
-
-  reset(hostId: number): void {
-    this.failures.delete(hostId);
-  }
-
-  cleanup(): void {
-    const maxAge = 60 * 60 * 1000;
-    const now = Date.now();
-
-    for (const [hostId, record] of this.failures.entries()) {
-      if (!record.permanent && now - record.lastFailure > maxAge) {
-        this.failures.delete(hostId);
-      }
-    }
-  }
-}
-
-class PollingBackoff {
-  private failures = new Map<number, { count: number; nextRetry: number }>();
-  private baseDelay = 30000;
-  private maxDelay = 600000;
-  private maxRetries = 5;
-
-  recordFailure(hostId: number): void {
-    const existing = this.failures.get(hostId) || { count: 0, nextRetry: 0 };
-    const delay = Math.min(
-      this.baseDelay * Math.pow(2, existing.count),
-      this.maxDelay,
-    );
-    this.failures.set(hostId, {
-      count: existing.count + 1,
-      nextRetry: Date.now() + delay,
-    });
-  }
-
-  shouldSkip(hostId: number): boolean {
-    const backoff = this.failures.get(hostId);
-    if (!backoff) return false;
-
-    if (backoff.count >= this.maxRetries) {
-      return true;
-    }
-
-    return Date.now() < backoff.nextRetry;
-  }
-
-  getBackoffInfo(hostId: number): string | null {
-    const backoff = this.failures.get(hostId);
-    if (!backoff) return null;
-
-    if (backoff.count >= this.maxRetries) {
-      return `Max retries exceeded (${backoff.count} failures) - polling suspended`;
-    }
-
-    const remainingMs = backoff.nextRetry - Date.now();
-    if (remainingMs > 0) {
-      const remainingSec = Math.ceil(remainingMs / 1000);
-      return `Retry in ${remainingSec}s (attempt ${backoff.count}/${this.maxRetries})`;
-    }
-
-    return null;
-  }
-
-  reset(hostId: number): void {
-    this.failures.delete(hostId);
-  }
-
-  cleanup(): void {
-    const maxAge = 60 * 60 * 1000;
-    const now = Date.now();
-
-    for (const [hostId, backoff] of this.failures.entries()) {
-      if (backoff.count < this.maxRetries && now - backoff.nextRetry > maxAge) {
-        this.failures.delete(hostId);
-      }
-    }
-  }
-}
-
-const requestQueue = new RequestQueue();
-const metricsCache = new MetricsCache();
-const authFailureTracker = new AuthFailureTracker();
-const pollingBackoff = new PollingBackoff();
 const authManager = AuthManager.getInstance();
 const permissionManager = PermissionManager.getInstance();
